Revamp of texture importing completed
many improvements in folder and material importers, had to solve some edgy case having to do with texture being both standalone reosurce and model related. Due to this a warning is printed saying s1 tried to add the same resource twice in a folder. (of course it is not added)