Solved bug with normal textures save/load (& found another!)
Color & Shininess of materials are not saved, only when changing them before dropping a texture (there's already an issue, #186)

This is because materialImporter's Save() function is called after dropping a texture, but not when changing color or shininess.

In addition, the Save() function is not called when saving scene, so unless you drop a texture (when starting engine, too, but that doesn't solves this xD), material it seems to never be saved.
Happens equal with the Load() function.